requirements.txt
astreamlit
pandas
aiohttp
beautifulsoup4
lxml
matplotlib